requirements.txt
requests==2.31.0
python-dotenv==1.0.0
tweepy==4.14.0
praw==7.7.1
pyspark==3.4.1
pymongo==4.5.0
pandas==2.0.3
numpy==1.24.3